Fix the programming failure that occurred on 128MB
The NuWriter firmware utilizes DRAM address 0x8b000000
as the temporary buffer for the download image to be programmed.
While this is suitable for DRAM sizes greater than or equal to
256MB, it is invalid for a 128MB DRAM size. In the case of 128MB
DRAM, address 0x8b000000 is actually mapped to DRAM address
0x83000000, which is not as expected and results in unexpected
errors caused by data overwriting.

This patch updates the address of the image temporary buffer
from 0x8b000000 to 0x87e00000 to resolve this issue.

Change-Id: Ib00a40bd0e2e4b7d4860a6682f4226eacfbda0b9