NAND ONFI issue workaround fix for B version chip
To address the misjudgment of NAND ONFI by the IBR in MA35 Series A
version ICs, the NuWriter firmware (xusb) included a compatibility
workaround.

Since the IBR in MA35 B version ICs has already fixed the ONFI
misjudgment issue, the workaround in NuWriter is no longer needed.

To handle both cases correctly, the firmware now detects whether
the IC is an A or B version. If it is an A version, the workaround
remains applied; if it is a B version, the workaround is skipped.

Change-Id: Ie700bf9a433fd34d05c6740f6685cd60681cb2df